refactor: decompose cron, channel_history, cli_init into packages (MR B5)
Decompose three monolithic files into clean packages:

Cron builtin:
- builtins/tools/cron/__init__.py     CronToolBuiltin facade
- builtins/tools/cron/models.py       ScheduleAtInput, ScheduleEveryInput, CancelScheduledInput
- builtins/tools/cron/tools.py        4 tool factory functions
- builtins/tools/cron/formatting.py   _parse_timestamp, _format_interval, _format_time_until
- builtins/tools/cron/scheduler_bridge.py  set_scheduler, _add_to_live_scheduler, _remove_from_live_scheduler

Channel history builtin:
- builtins/tools/channel_history/__init__.py  ChannelHistoryToolBuiltin facade
- builtins/tools/channel_history/models.py    BrowseChannelHistoryInput
- builtins/tools/channel_history/resolver.py  _resolve_adapter, _resolve_channel_id
- builtins/tools/channel_history/formatter.py _format_history_output, _format_timestamp, _build_filter_qualifier
- builtins/tools/channel_history/browser.py   browse_sync, _browse_async, _create_browse_tool

CLI init:
- cli_init/__init__.py   Re-export dispatch_command
- cli_init/templates.py  Template constants, _build_agent_yaml, _parse_model_string
- cli_init/scaffolder.py _validate_workspace_name, _create_workspace, _scaffold_default_profiles
- cli_init/commands.py   _handle_init, _handle_list, dispatch_command

Tests split alongside sources:
- tests/builtins/cron/test_tools.py + test_store.py
- tests/builtins/channel_history/test_channel_history_builtin.py + test_browser.py + test_formatter.py + test_resolver.py + test_wiring.py
- tests/cli_init/test_templates.py + test_scaffolder.py + test_commands.py

All 2944 tests passing. Ruff clean.
